+++ src/handlers/default.events/enable_plugin.js
@@ -4,10 +4,12 @@ const eventNames = require("../eventNames");
 const fs = require("node:fs");
 
 module.exports = ({ io, data }) => {
+	if(!fs.existsSync(path.resolve(`./plugins/${data}`))) return;
 	fs.renameSync(
-		path.resolve(`./plugins/${data}.Freedeck.disabled`),
-		path.resolve(`./plugins/${data}.Freedeck`),
+		path.resolve(`./plugins/${data}`),
+		path.resolve(`./plugins/${data.split(".disabled")[0]}`),
 	);
-	plugins.reload();
+	plugins._disabled = plugins._disabled.filter((x) => x !== data);
+	plugins.load(data.split(".disabled")[0]);
 	io.emit(eventNames.default.reload);
 };
